fade out exiting task bar buttons and tabs

+++ kernel/files.cpp
@@ -84,7 +84,7 @@ struct {
 	bool shutdown;
 
 	volatile uint64_t totalHandleCount;
-	volatile uint64_t fileSystemsUnmounting;
+	volatile uintptr_t fileSystemsUnmounting;
 	KEvent fileSystemUnmounted;
 } fs = {
 	.fileSystemUnmounted = { .autoReset = true },
